refactor: Entity 설계
