+++ src/main/java/racingcar/dao/GameDao.java
@@ -5,7 +5,6 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import racingcar.dto.GameDto;
 
 @Repository
 public class GameDao {
@@ -20,8 +19,8 @@ public class GameDao {
                 .usingGeneratedKeyColumns("game_id");
     }
 
-    public int save(GameDto gameDto) {
-        SqlParameterSource sqlParameterSource = new BeanPropertySqlParameterSource(gameDto);
+    public int save(GameEntity gameEntity) {
+        SqlParameterSource sqlParameterSource = new BeanPropertySqlParameterSource(gameEntity);
         return insertActor.executeAndReturnKey(sqlParameterSource).intValue();
     }
 }
